requirements.txt
flask
youtube_dl
pillow
requests
dhooks
